+++ src/core/store/wallet/actions.js
@@ -10,12 +10,16 @@ import { EventBus } from '@/core/plugins/eventBus';
 const removeWallet = function ({ commit, state }) {
   if (
     state.identifier === WALLET_TYPES.WALLET_CONNECT ||
-    state.identifier === WALLET_TYPES.WALLET_LINK ||
-    state.identifier === WALLET_TYPES.MEW_CONNECT
+    state.identifier === WALLET_TYPES.WALLET_LINK
   ) {
     const connection = state.instance.getConnection();
-    if (connection && connection.disconnect) {
-      connection.disconnect();
+    if (connection) {
+      if (connection.disconnect) {
+        connection.disconnect();
+      }
+      if (connection.killSession) {
+        connection.killSession();
+      }
     }
   }
   commit('REMOVE_WALLET');
@@ -50,7 +54,7 @@ const setWeb3Instance = function (
     rootState.global.currentNetwork.port
       ? ':' + rootState.global.currentNetwork.port
       : ''
-  }${hostUrl.pathname}`;
+  }${hostUrl.pathname ? hostUrl.pathname : ''}`;
   rootState.global.currentNetwork.username !== '' &&
   rootState.global.currentNetwork.password !== ''
     ? (options['headers'] = {
@@ -134,6 +138,9 @@ const setOfflineApp = function ({ commit }, val) {
 const setLedgerApp = function ({ commit }, val) {
   commit('SET_LEDGER_APP', val);
 };
+const setSwapRates = function ({ commit }, val) {
+  commit('SET_SWAP_RATES', val);
+};
 
 export default {
   removeWallet,
@@ -145,5 +152,6 @@ export default {
   setOwnedDomains,
   setTokens,
   setOfflineApp,
-  setLedgerApp
+  setLedgerApp,
+  setSwapRates
 };