+++ src/modules/confirmation/ModuleConfirmation.vue
@@ -751,7 +751,7 @@ export default {
           .catch(err => {
             if (_this.isSwap && idx + 1 === _arr.length) {
               if (rejectedError(err.message)) {
-                _this.trackSwap('swapTxCancelled');
+                _this.trackSwap('swapTxRejected');
               } else {
                 _this.emitSwapTxFail(err);
               }
@@ -844,7 +844,7 @@ export default {
           .catch(e => {
             if (this.isSwap) {
               if (rejectedError(e.message)) {
-                this.trackSwap('swapTxCancelled');
+                this.trackSwap('swapTxRejected');
               } else {
                 this.emitSwapTxFail(e);
               }
@@ -919,7 +919,7 @@ export default {
               .catch(e => {
                 if (this.isSwap) {
                   if (rejectedError(e.message)) {
-                    this.trackSwap('swapTxCancelled');
+                    this.trackSwap('swapTxRejected');
                     throw new Error(e.message);
                   }